# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SolaBridge API",
        "description": "Documentação oficial da API do SolaBridge ERP (Motor Contábil).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/servicos": {
            "get": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Lista (paginada) os serviços do tenant, com busca por nome/código",
                "operationId": "716bb4129123ce67b7baba7785ae7ac4",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Busca por nome ou código",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de serviços."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Cadastra um serviço para o tenant",
                "operationId": "310641dcdd4b36dce76883fe6ec1b8a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "descricao",
                                    "valor_servico",
                                    "codigo",
                                    "codigo_tributacao_nacional",
                                    "codigo_ibge_municipio"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Consultoria Contábil"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Consultoria contábil mensal"
                                    },
                                    "valor_servico": {
                                        "type": "number",
                                        "example": 500
                                    },
                                    "codigo": {
                                        "type": "string",
                                        "example": "SRV-001"
                                    },
                                    "codigo_tributacao_nacional": {
                                        "type": "string",
                                        "example": "010701"
                                    },
                                    "codigo_ibge_municipio": {
                                        "type": "string",
                                        "example": "3550308"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Serviço criado."
                    },
                    "422": {
                        "description": "Erro de validação."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/servicos/{id}": {
            "delete": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Exclusão lógica (soft delete) de um serviço",
                "operationId": "1cbfb2a065ac384a9b14432431f2bb69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Serviço inativado (soft delete)."
                    },
                    "404": {
                        "description": "Serviço não encontrado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Atualização parcial de um serviço",
                "operationId": "23826ae7ee80f386ba1940e7563cb349",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Consultoria Contábil Premium"
                                    },
                                    "valor_servico": {
                                        "type": "number",
                                        "example": 750
                                    },
                                    "aliquota_iss": {
                                        "type": "number",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Serviço atualizado."
                    },
                    "404": {
                        "description": "Serviço não encontrado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token pessoal (Sanctum) retornado pelo endpoint /login.",
                "scheme": "bearer"
            }
        }
    },
...
}